requirements.txt
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
et-xmlfile==1.1.0
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
openpyxl==3.1.2
pdfkit==1.0.0
pkg_resources==0.0.0
python-dotenv==1.0.1
requests==2.31.0
resend==0.8.0
soupsieve==2.5
urllib3==1.26.18
